+++ app/src/components/datasets/DatasetContentTabs/DatasetContentTabs.tsx
@@ -30,6 +30,8 @@ const tabs = [
   },
 ];
 
-const DatasetContentTabs = () => <ContentTabs tabs={tabs} headerProps={{ px: 8 }} />;
+const DatasetContentTabs = () => (
+  <ContentTabs tabs={tabs} headerProps={{ px: 8, position: "sticky", left: 0, right: 0 }} />
+);
 
 export default DatasetContentTabs;
