+++ app/src/pages/datasets/[id]/[tab].tsx
@@ -51,8 +51,8 @@ export default function Dataset() {
   }
 
   return (
-    <AppShell title={dataset.data?.name}>
-      <VStack h="full">
+    <AppShell title={dataset.data?.name} containerProps={{ position: "relative" }}>
+      <VStack position="sticky" left={0} right={0} w="full">
         <BetaBanner />
         <PageHeaderContainer>
           <Breadcrumb>
@@ -84,8 +84,8 @@ export default function Dataset() {
             </BreadcrumbItem>
           </Breadcrumb>
         </PageHeaderContainer>
-        <DatasetContentTabs />
       </VStack>
+      <DatasetContentTabs />
       <FileUploadsCard />
     </AppShell>
   );
